requirements.txt
selenium==4.15.2
pillow==10.1.0
psutil==5.9.6